Fix default args being read-only
This happened when no .ini file was supplied. Then the default args were
supposed to be used, but this was never tested. Since it was a string
literal it ended up in the read-only area of the program. Later, trying
to parse the args, strtok_r tried to change this read-only value. The
solution was to make the default arguments constant, and copy them into
a writable variable.